+++ src/DefinitionProvider.ts
@@ -19,7 +19,11 @@ export class DefinitionProvider implements vscode.DefinitionProvider {
      * @param token 
      */
     public provideDefinition(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken): Thenable<vscode.Location[]> {
-        // Check for 'include "..."'
+        const settings = vscode.workspace.getConfiguration('asm-code-lens');
+        if(settings.enableGotoDefinition == false)
+            return undefined;
+
+            // Check for 'include "..."'
         const lineContents = document.lineAt(position.line).text;
         const match = regexInclude().exec(lineContents);
         if(match) {