+++ src/HoverProvider.ts
@@ -19,9 +19,8 @@ export class HoverProvider implements vscode.HoverProvider {
      * @param token 
      */
     public provideHover(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken): Thenable<vscode.Hover> {
-        let settings = vscode.workspace.getConfiguration('asm-code-lens');
-        if(settings.enableHoverProvider == undefined
-            || settings.enableHoverProvider)
+        const settings = vscode.workspace.getConfiguration('asm-code-lens');
+        if(settings.enableHovering != false)
             return this.search(document, position);
         else
             return undefined;